+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.23.7
 
 - Use only `dart` commands, Flutter SDK will be used when `FLUTTER_ROOT` is set.
+- Option to use a precompiled `dartdoc` binary.
 
 ## 0.23.6